# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "DotAAS Part 2 | HTTP/REST | Entire Interface Collection",
    "description" : "The entire interface collection as part of Details of the Asset Administration Shell Part 2",
    "termsOfService" : "https://github.com/admin-shell-io/aas-specs",
    "contact" : {
      "name" : "Michael Hoffmeister, Torben Miny, Andreas Orzelski, Manuel Sauer, Constantin Ziesche"
    },
    "version" : "V1.0RC01"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/concept-descriptions" : {
      "get" : {
        "tags" : [ "Concept Description Repository Interface" ],
        "summary" : "Returns all Concept Descriptions",
        "operationId" : "GetAllConceptDescriptions",
        "parameters" : [ {
          "name" : "idShort",
          "in" : "query",
          "description" : "The Concept Description’s IdShort",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "isCaseOf",
          "in" : "query",
          "description" : "IsCaseOf reference (BASE64-URL-encoded)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dataSpecificationRef",
          "in" : "query",
          "description" : "DataSpecification reference (BASE64-URL-encoded)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Requested Concept Descriptions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ConceptDescription"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GetAllConceptDescriptions/1/0/RC02", "https://admin-shell.io/aas/API/GetAllConceptDescriptionsByIdShort/1/0/RC02", "https://admin-shell.io/aas/API/GetAllConceptDescriptionsByIsCaseOf/1/0/RC02", "https://admin-shell.io/aas/API/GetAllConceptDescriptionsByDataSpecificationReference/1/0/RC02" ]
      },
      "post" : {
        "tags" : [ "Concept Description Repository Interface" ],
        "summary" : "Creates a new Concept Description",
        "operationId" : "PostConceptDescription",
        "requestBody" : {
          "description" : "Concept Description object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConceptDescription"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Concept Description created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConceptDescription"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/PostConceptDescription/1/0/RC02" ]
      }
    },
    "/concept-descriptions/{cdIdentifier}" : {
      "get" : {
        "tags" : [ "Concept Description Repository Interface" ],
        "summary" : "Returns a specific Concept Description",
        "operationId" : "GetConceptDescriptionById",
        "parameters" : [ {
          "name" : "cdIdentifier",
          "in" : "path",
          "description" : "The Concept Description’s unique id (BASE64-URL-encoded)",
          "required" : true,
          "style" : "simple",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Requested Concept Description",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConceptDescription"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GetConceptDescriptionById/1/0/RC02" ]
      },
      "put" : {
        "tags" : [ "Concept Description Repository Interface" ],
        "summary" : "Updates an existing Concept Description",
        "operationId" : "PutConceptDescriptionById",
        "parameters" : [ {
          "name" : "cdIdentifier",
          "in" : "path",
          "description" : "The Concept Description’s unique id (BASE64-URL-encoded)",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Concept Description object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConceptDescription"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Concept Description updated successfully"
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/PutConceptDescriptionById/1/0/RC02" ]
      },
      "delete" : {
        "tags" : [ "Concept Description Repository Interface" ],
        "summary" : "Deletes a Concept Description",
        "operationId" : "DeleteConceptDescriptionById",
        "parameters" : [ {
          "name" : "cdIdentifier",
          "in" : "path",
          "description" : "The Concept Description’s unique id (BASE64-URL-encoded)",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Concept Description deleted successfully"
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/DeleteConceptDescriptionById/1/0/RC02" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "HasSemantics" : {
        "type" : "object",
        "properties" : {
          "semanticId" : {
            "$ref" : "#/components/schemas/Reference"
          }
        }
      },
      "Reference" : {
        "type" : "object",
        "oneOf" : [ {
          "$ref" : "#/components/schemas/GlobalReference"
        }, {
          "$ref" : "#/components/schemas/ModelReference"
        } ]
      },
      "GlobalReference" : {
        "required" : [ "value" ],
        "properties" : {
          "value" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          }
        }
      },
      "Identifier" : {
        "type" : "string"
      },
      "ModelReference" : {
        "required" : [ "keys" ],
        "properties" : {
          "referredSemanticId" : {
            "$ref" : "#/components/schemas/GlobalReference"
          },
          "keys" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Key"
            }
          }
        }
      },
      "Key" : {
        "required" : [ "type", "value" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/KeyElements"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
      "KeyElements" : {
        "type" : "string",
        "enum" : [ "AssetAdministrationShell", "AccessPermissionRule", "ConceptDescription", "Submodel", "AnnotatedRelationshipElement", "BasicEvent", "Blob", "Capability", "DataElement", "File", "Entity", "Event", "MultiLanguageProperty", "Operation", "Property", "Range", "ReferenceElement", "RelationshipElement", "SubmodelElement", "SubmodelElementList", "SubmodelElementStruct", "View", "FragmentReference" ]
      },
...
      "Identifiable" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Referable"
        }, {
          "required" : [ "identification" ],
          "properties" : {
            "administration" : {
              "$ref" : "#/components/schemas/AdministrativeInformation"
            },
            "identification" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          }
        } ]
      },
      "Referable" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasExtensions"
        }, {
          "required" : [ "idShort", "modelType" ],
          "properties" : {
            "category" : {
              "type" : "string"
            },
            "description" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "displayName" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "idShort" : {
              "type" : "string"
            },
            "modelType" : {
              "$ref" : "#/components/schemas/ModelType"
            }
          }
        } ]
      },
      "HasExtensions" : {
        "type" : "object",
        "properties" : {
          "extensions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          }
        }
      },
      "Extension" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "required" : [ "name" ],
          "properties" : {
            "name" : {
              "type" : "string"
            },
            "refersTo" : {
              "$ref" : "#/components/schemas/Reference"
            },
            "value" : {
              "type" : "string"
            },
            "valueType" : {
              "$ref" : "#/components/schemas/ValueTypeEnum"
            }
          }
        } ]
      },
      "ValueTypeEnum" : {
        "type" : "string",
        "enum" : [ "anyUri", "base64Binary", "boolean", "date", "dateTime", "dateTimeStamp", "decimal", "integer", "long", "int", "short", "byte", "nonNegativeInteger", "positiveInteger", "unsignedLong", "unsignedInt", "unsignedShort", "unsignedByte", "nonPositiveInteger", "negativeInteger", "double", "duration", "dayTimeDuration", "yearMonthDuration", "float", "gDay", "gMonth", "gMonthDay", "gYear", "gYearMonth", "hexBinary", "NOTATION", "QName", "string", "normalizedString", "token", "language", "Name", "NCName", "ENTITY", "ID", "IDREF", "NMTOKEN", "time" ]
      },
      "LangString" : {
        "required" : [ "language", "text" ],
        "type" : "object",
        "properties" : {
          "language" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        },
        "example" : {
          "language" : "language",
          "text" : "text"
        }
      },
      "ModelType" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "$ref" : "#/components/schemas/ModelTypes"
          }
        }
      },
      "ModelTypes" : {
        "type" : "string",
        "enum" : [ "AssetAdministrationShell", "ConceptDescription", "Submodel", "AnnotatedRelationshipElement", "BasicEvent", "Blob", "Capability", "DataElement", "File", "Entity", "Event", "ModelReference", "MultiLanguageProperty", "Operation", "Property", "Range", "ReferenceElement", "RelationshipElement", "SubmodelElement", "SubmodelElementList", "SubmodelElementStruct", "View", "GlobalReference", "FragmentReference", "Constraint", "Formula", "Qualifier" ]
      },
      "AdministrativeInformation" : {
        "type" : "object",
        "properties" : {
          "revision" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        },
        "example" : {
          "version" : "version",
          "revision" : "revision"
        }
      },
      "HasDataSpecification" : {
        "type" : "object",
        "properties" : {
          "embeddedDataSpecifications" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EmbeddedDataSpecification"
            }
          }
        }
      },
      "EmbeddedDataSpecification" : {
        "required" : [ "dataSpecification", "dataSpecificationContent" ],
        "type" : "object",
        "properties" : {
          "dataSpecification" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "dataSpecificationContent" : {
            "$ref" : "#/components/schemas/DataSpecificationContent"
          }
        }
      },
      "DataSpecificationContent" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/DataSpecificationIEC61360Content"
        }, {
          "$ref" : "#/components/schemas/DataSpecificationPhysicalUnitContent"
        } ]
      },
      "DataSpecificationIEC61360Content" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/ValueObject"
        }, {
          "required" : [ "preferredName" ],
          "type" : "object",
          "properties" : {
            "dataType" : {
              "type" : "string",
              "enum" : [ "DATE", "STRING", "STRING_TRANSLATABLE", "REAL_MEASURE", "REAL_COUNT", "REAL_CURRENCY", "BOOLEAN", "URL", "RATIONAL", "RATIONAL_MEASURE", "TIME", "TIMESTAMP", "INTEGER_COUNT", "INTEGER_MEASURE", "INTEGER_CURRENCY" ]
            },
            "definition" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "levelType" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LevelType"
              }
            },
            "preferredName" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "shortName" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "sourceOfDefinition" : {
              "type" : "string"
            },
            "symbol" : {
              "type" : "string"
            },
            "unit" : {
              "type" : "string"
            },
            "unitId" : {
              "$ref" : "#/components/schemas/Reference"
            },
            "valueFormat" : {
              "type" : "string"
            },
            "valueList" : {
              "$ref" : "#/components/schemas/ValueList"
            }
          }
        } ]
      },
      "ValueObject" : {
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          },
          "valueId" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "valueType" : {
            "$ref" : "#/components/schemas/ValueTypeEnum"
          }
        }
      },
      "LevelType" : {
        "type" : "string",
        "enum" : [ "Min", "Max", "Nom", "Typ" ]
      },
      "ValueList" : {
        "required" : [ "valueReferencePairTypes" ],
        "type" : "object",
        "properties" : {
          "valueReferencePairTypes" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ValueReferencePairType"
            }
          }
        }
      },
      "ValueReferencePairType" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/ValueObject"
        } ]
      },
      "DataSpecificationPhysicalUnitContent" : {
        "required" : [ "definition", "unitName", "unitSymbol" ],
        "type" : "object",
        "properties" : {
          "conversionFactor" : {
            "type" : "string"
          },
          "definition" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LangString"
            }
          },
          "dinNotation" : {
            "type" : "string"
          },
          "eceCode" : {
            "type" : "string"
          },
          "eceName" : {
            "type" : "string"
          },
          "nistName" : {
            "type" : "string"
          },
          "registrationAuthorityId" : {
            "type" : "string"
          },
          "siName" : {
            "type" : "string"
          },
          "siNotation" : {
            "type" : "string"
          },
          "sourceOfDefinition" : {
            "type" : "string"
          },
          "supplier" : {
            "type" : "string"
          },
          "unitName" : {
            "type" : "string"
          },
          "unitSymbol" : {
            "type" : "string"
          }
        }
      },
...
      "ConceptDescription" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "properties" : {
            "isCaseOf" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Reference"
              }
            }
          }
        } ]
      },
...
    }
  }
}